chore(translation.json): add "identify-location" key to translation files in multiple languages
The "identify-location" key was added to the translation files in Arabic, English, Finnish, and Italian languages. This key is used to translate the phrase "Identify location" in the respective languages. This change was made to improve the localization of the application and provide a consistent user experience across different languages.